refactor: front-end refactor components
