+++ app/components/Attachment/attachment.scss
@@ -1,0 +1,37 @@
+/*upload file*/
+.file_li{
+  list-style: none;
+  display: flex;
+  margin-top: 3px;
+  margin-bottom: 3px;
+  align-self: flex-end;
+}
+.file-icon-container img{
+  height: 25px;
+  display: inline-block;
+  margin-right: 10px;
+}
+.file_option{
+  margin-left: 5px;
+  margin-top: 5px;
+  height: 20px;
+}
+
+.container{
+  width: 100%;
+}
+.upload_file_zone{
+  flex: 1;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  padding:20px;
+  border-width: 1.5px;
+  border-radius: 2px;
+  border-color: #69cb43;
+  border-style: dashed;
+  background-color: #fafafa;
+  color: #69CB43;
+  outline: none;
+  transition: border .24s ease-in-out;
+}
