+++ app/components/Block/ImageBlock.jsx
@@ -1,17 +1,16 @@
 import React, { useState } from 'react';
-import Upload from '../Upload';
 import BlockTitle from "../BlockTitle/BlockTitle";
-import BlockDescription from "../BlockDescription";
+import BlockDescription from "../BlockDescription/BlockDescription";
 
+import './block.scss'
 // Icons
-import BlockIcon from "../../asset/collection/quote-right.svg";
+import BlockIcon from "../../asset/collection/images.svg";
 import TrashIcon from "../../asset/collection/trash-alt.svg";
 import AngleIcon from "../../asset/collection/angle-up.svg";
 import MarkIcon from "../../asset/collection/bookmark.svg";
 import MarkFullIcon from "../../asset/collection/bookmark-full.svg";
 
-
-export default function TextBlock(props) {
+export default function ImageBlock(props) {
     const scaleid = "scale_" + props.block.timestamp;
     const checkid = "check_" + props.block.timestamp;
 
@@ -26,11 +25,15 @@ export default function TextBlock(props) {
             document.getElementById(scaleid).classList.remove("rotate-close");
             document.getElementById(scaleid).classList.toggle("rotate-open");
         }
-        setScaling(!scaling);
+        setScaling(!scaling)
+    }
+
+    let showImageWindow = () => {
+        window.open(props.block.paths[0]);
     }
 
     return (
-        <div id={props.block.timestamp} className="textBlock blockContent">
+        <div id={props.block.timestamp} className="picBlock blockContent">
             <div className="borderLine"></div>
             <BlockTitle className="blockTitle" time={props.block.timestamp} onChangeTitle={props.handleTitle} title={props.block.title} />
             <button className="iconBtn removeBtn" onClick={props.delBlock.bind(this, props.block.timestamp)}><img src={TrashIcon}></img></button>
@@ -46,17 +49,12 @@ export default function TextBlock(props) {
             <button className="iconBtn scaleBtn" onClick={handleScaling}><img src={AngleIcon} id={scaleid}></img></button>
             {scaling &&
                 <div className="blockMain">
+                    <img src={props.block.paths[0]} onDoubleClick={showImageWindow} />
                     <BlockDescription
                         description={props.block.description}
                         addDescription={props.addDescription}
                         time={props.block.timestamp}
                         handleLinker={props.handleLinker}
-                    />
-                    <Upload
-                        paths={props.block.paths}
-                        time={props.block.timestamp}
-                        addFile={props.addFile}
-                        delFile={props.delFile}
                     />
                 </div>
             }
