+++ app/components/Block/block.scss
@@ -1,0 +1,133 @@
+.blockContent{
+  font-size: 16px;
+  color: #143938;
+  margin:0 10% 10px 10%;
+
+  display: grid;
+  grid-template-rows: 50px 30px auto auto;
+  grid-template-columns:50px 40px 80px 40px auto 50px 50px;
+  grid-template-areas:
+    "date . icon       scale   title   mark     remove "
+    ".    . time       content content content content "
+    ".    . checkbox   content content content content "
+    ".    .  .         content content content content ";
+}
+.borderLine{
+  grid-column:4 /-1;
+  grid-row:  1 / -1;
+}
+.blockContent:hover .borderLine{
+  border: 1.8px solid #2b7a4f;
+  border-radius: 5px;
+  box-shadow:0 3px 4px 3px rgba(0,0,0,0.25);
+  transition: 0.3s;
+}
+
+/*Button */
+.iconBtn{
+  outline: none !important;
+  outline-offset: none !important;
+  background:none;
+  border: none;
+  align-self: center;
+}
+.iconBtn img{
+  width: 20px;
+  cursor: pointer;  
+}
+.markBtn{
+  grid-area: mark;
+  text-align: center;
+}
+.removeBtn{
+  grid-area: remove;
+}
+.scaleBtn{
+  grid-area: scale;
+  padding-left: 20px;
+}
+/*button end */
+
+.checkContainer{
+  grid-area: checkbox;
+  align-self: center;
+  justify-self: center;
+  display: none;
+}
+.check{
+  opacity: 0;
+}
+
+.checkmark{
+  position: relative;
+  left: -15px;
+}
+
+.checkmark::before{
+  content: "";
+  display: inline-block;
+  height: 15px;
+  width: 15px;
+  border-radius: 100%;
+  border: 1.5px solid #35c189;
+  cursor: pointer;  
+}
+.checkmark::after {
+  color: #35c189;
+  content: "";
+  display: inline-block;
+  height: 6px;
+  width: 9px;
+  left: 3px;
+  top: 4px;
+  border-left: 2px solid;
+  border-bottom: 2px solid;
+  transform: rotate(-45deg);
+}
+.checkmark::before, .checkmark::after{
+  position: absolute;
+}
+.check +label::after{
+  content: none;
+}
+.check:checked + label::after{
+  content:"";
+}
+
+.timeINFO{
+  text-align: center;
+  align-self: left;
+  color: #35c189;
+}
+.date{
+  grid-area: date;
+  font-size: 16px;
+  font-weight: 800;
+  padding-top: 15px;
+  text-align: right;
+}
+.time{
+  grid-area: time;
+  font-size: 16px;
+}
+
+.blockIcon{
+  grid-area: icon;
+  /* border: black solid 1px; */
+  text-align: center;
+  align-self: center;
+}
+.blockIcon img{
+  height: 30px;
+}
+
+.blockMain{
+  grid-area: content;
+  padding: 20px 20px;
+  line-height: 30px;
+}
+
+.picBlock img{
+  max-width: 90%;
+  max-height: auto;
+}
