+++ app/components/BlockDescription/blockDescription.scss
@@ -1,0 +1,37 @@
+
+/* input */
+.inputContainer{
+  width: 100%;
+  margin: 20px 0;
+  border-radius: 4px;
+  text-align: left;
+  box-sizing: border-box;
+  word-wrap: break-word;
+}
+.inputField{
+  width:50%;
+  height: 24px;
+  border: 0;
+  border-radius: 4px;
+  background-color: #35c18927;
+  color: #35c189;
+  font-size: 16px;
+  word-wrap: break-word;
+}
+.inputField_title{
+  width:30%;
+  height: 60%;
+  border: 0;
+  border-radius: 4px;
+  background-color: rgba(189, 212, 203, 0.61);
+  color: #143938;
+  font-size: 22px;
+  word-wrap: break-word;
+  text-align: center;
+}
+.inputField_title:focus{
+  outline: none;
+}
+.descriptionInput::placeholder{
+  color:#1e525054;
+}
